+++ SuCoS.Generator/SuCosGenerator.cs
@@ -19,10 +19,19 @@ public class SuCosGenerator : ISourceGenerator
         var source = $@"
 namespace SuCoS.Models;
 
-public class SucosExt
+/// <summary>
+/// (Code generated from SuCoSGenerator.cs) Build/compilation metadata.
+/// </summary>
+public static class SucosExt
 {{
+    /// <summary>
+    /// Date and time in UTC.
+    /// </summary>
     public static DateTime BuildDate => new DateTime({DateTime.UtcNow.Ticks}, DateTimeKind.Utc);
 
+    /// <summary>
+    /// Date and time (expressed as Ticks) in UTC.
+    /// </summary>
     public static long BuildDateTicks => {DateTime.UtcNow.Ticks};
 }}
 ";